+++ vue/src/Components/Chat/ChatMessagesList.vue
@@ -1,9 +1,19 @@
 <template>
-  <ul class="ai-chat-conversation">
-    <ChatMessage v-for="(message, index) in messages" :message="message" :key="index"/>
-    <ChatLoading v-if="loading && !errored"/>
+  <div ref="conversationWrapper" class="ai-chat-conversation-wrapper">
+
+    <ul class="ai-chat-conversation" v-if="messages.length">
+      <ChatMessage
+        v-for="(message, index) in messages"
+        :message="message"
+        :key="index"
+        :ai="ai"
+        :primary-color="primaryColor"
+      />
+    </ul>
+
+    <ChatLoading v-if="loading && !errored" :ai="ai" :primary-color="primaryColor"/>
     <Alert v-if="errored" severity="danger">Ooops, AI have encountered an error.</Alert>
-  </ul>
+  </div>
 </template>
 
 <script lang="ts">
@@ -32,15 +42,29 @@ export default defineComponent({
       type: Array,
       default: () => [],
     },
+    ai: {
+      type: String,
+      required: true,
+    },
+    primaryColor: {
+      type: String,
+      default: '#3450a3',
+    },
+  },
+  methods: {
+    scrollDown() {
+      setTimeout(() => {
+        this.$refs.conversationWrapper.scrollTo(0, document.body.scrollHeight);
+      }, 1);
+    },
   },
 });
 </script>
 
 <style lang="less" scoped>
-.ai-chat-conversation {
-  padding-bottom: 50px;
-  padding-left: 0;
-  margin-bottom: 0;
+.ai-chat-conversation-wrapper {
+  flex-grow: 1;
+  width: 100%;
   margin-left: auto;
   margin-right: auto;
   max-width: 730px;
@@ -48,5 +72,18 @@ export default defineComponent({
   display: flex;
   flex-direction: column;
   gap: 1.5rem;
+  overflow-y: auto;
+  scrollbar-width: none;
+  padding: 1.5rem 0;
+  max-height: calc(100% - 91px);
+
+  .ai-chat-conversation {
+    padding-left: 0;
+    margin-bottom: 0;
+    list-style-type: none;
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+  }
 }
 </style>