+++ vue/src/Components/Insight/InsightTrigger.vue
@@ -5,10 +5,11 @@
     </button>
 
     <InsightOffcanvas
+      ref="offCanvas"
       :display-offcanvas="displayOffcanvas"
-      :markdown="markdown"
-      :loading="loading"
-      :errored="errored"
+      :report-id="reportId"
+      :ai="ai"
+      :primary-color="primaryColor"
       @close="onClose"/>
 
   </div>
@@ -16,7 +17,6 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import { AjaxHelper } from 'CoreHome';
 import InsightOffcanvas from './InsightOffcanvas.vue';
 import IconMagic from '../Icon/IconMagic.vue';
 
@@ -30,43 +30,26 @@ export default defineComponent({
       type: String,
       required: true,
     },
+    ai: {
+      type: String,
+      required: true,
+    },
+    primaryColor: {
+      type: String,
+      default: '#3450a3',
+    },
   },
   data() {
     return {
       displayOffcanvas: false,
-      loading: false,
-      errored: false,
-      markdown: '',
     };
   },
   methods: {
     onClick() {
       this.displayOffcanvas = !this.displayOffcanvas;
-      this.loading = true;
-      AjaxHelper
-        .fetch({
-          method: 'ChatGPT.getInsights',
-          reportId: this.reportId,
-        })
-        .then((response) => {
-          if (response.choices && response.choices.length) {
-            this.markdown = response.choices[0].message.content;
-          } else {
-            this.errored = true;
-            console.error(response.message);
-            this.markdown = response.message;
-          }
-        })
-        .catch((error) => {
-          console.error(error);
-          this.errored = true;
-        })
-        .finally(() => {
-          this.loading = false;
-        });
+      this.$refs.offCanvas.onSubmit();
     },
     onClose() {
-      this.errored = false;
       this.displayOffcanvas = false;
     },
   },
@@ -75,22 +58,24 @@ export default defineComponent({
 
 <style lang="less" scoped>
 .ai-chat-insight-trigger-button {
+  background-color: transparent;
+  padding: 0;
   cursor: pointer;
   float: right;
-  padding: 4px;
-  border: 1px solid #00A67E;
-  background-color: #fff;
-  color: #00A67E;
+  border: 1px solid v-bind(primaryColor);
+  color: v-bind(primaryColor);
   display: flex;
   justify-content: center;
   align-items: center;
   border-radius: 5px;
   transition: .2s ease all;
+  width: 1.5rem;
+  height: 1.5rem;
 
   &:hover,
   &:focus {
     color: #fff;
-    background-color: #00A67E;
+    background-color: v-bind(primaryColor);
   }
 
   svg {