+++ vue/src/Pages/ChatIndexPage.vue
@@ -1,22 +1,90 @@
 <template>
   <div class="ai-chat-page-wrapper">
-    <Chat/>
+    <Chat
+      ref="chat"
+      :loading="loading"
+      :errored="errored"
+      :messages="messages"
+      :ai="ai"
+      :primary-color="primaryColor"
+      @prompt="onSubmit"
+    />
   </div>
 </template>
 
 <script lang="ts">
+import { AjaxHelper } from 'CoreHome';
 import { defineComponent } from 'vue';
 import Chat from '../Components/Chat/Chat.vue';
+
+interface MessageState {
+  role: string,
+  content: string,
+}
 
 export default defineComponent({
   components: {
     Chat,
   },
+  props: {
+    ai: {
+      type: String,
+      required: true,
+    },
+    primaryColor: {
+      type: String,
+      default: '#3450a3',
+    },
+  },
+  data() {
+    return {
+      loading: false,
+      errored: false,
+      messages: [],
+    };
+  },
+  methods: {
+    onSubmit(userPrompt) {
+      this.loading = true;
+      if (userPrompt) {
+        this.messages.push(userPrompt);
+      }
+      this.$refs.chat.scrollDown();
+      AjaxHelper
+        .fetch({
+          method: 'ChatGPT.getResponse',
+          messages: this.messages,
+        })
+        .then((response) => {
+          if (response.choices && response.choices.length > 0) {
+            this.messages.push(response.choices[0].message);
+          }
+        })
+        .catch(() => {
+          this.errored = true;
+        })
+        .finally(() => {
+          this.loading = false;
+          this.$refs.chat.scrollDown();
+        });
+    },
+  },
 });
 </script>
 
-<style lang="less" scoped>
+<style lang="less">
 .ai-chat-page-wrapper {
+  height: calc(100vh - 200px);
+  display: flex;
+  flex-direction: column;
 
+  .ai-chat-conversation-wrapper {
+    //max-height: none;
+  }
+
+  .ai-chat-form-wrapper {
+    position: fixed;
+    bottom: 20px;
+  }
 }
 </style>